Re-organizing code into monorepo
