+++ next-visual/cypress.config.ts
