+++ next-visual/cypress/component/NextVisual.cy.tsx
