+++ next-visual/cypress/support/component.ts
