+++ next-visual/cypress/tsconfig.json
