+++ next-visual/src/lib/placeholder.ts
