+++ next-visual/src/types/nextVisualTypes.ts
