+++ package.json
@@ -1,37 +1,12 @@
 {
-  "name": "next-visual",
-  "version": "0.1.3",
-  "description": "Image and video renderer for Next.js projects",
-  "repository": "git@github.com:BKWLD/next-visual.git",
-  "author": "Robert Reinhard <robert.reinhard@bukwild.com>",
+  "name": "react-visual-components",
+  "version": "0.0.1",
+  "description": "Image and video renderers for React projects",
+  "author": "Bukwild <info@bukwild.com>",
   "license": "MIT",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "files": [
-    "dist/**/*"
-  ],
-  "scripts": {
-    "build": "tsc",
-    "test": "cypress run --component",
-    "prepare": "yarn build",
-    "preversion": "yarn test",
-    "version": "git add -A",
-    "postversion": "git push --follow-tags"
-  },
-  "dependencies": {
-    "react-intersection-observer": "^9"
-  },
-  "peerDependencies": {
-    "next": "^13",
-    "react": "^18"
-  },
-  "devDependencies": {
-    "@types/react": "^18.2.18",
-    "@types/react-dom": "^18.2.7",
-    "cypress": "^12.17.3",
-    "next": "^13",
-    "react": "^18",
-    "react-dom": "^18.2.0",
-    "typescript": "^5.1.6"
-  }
+  "private": true,
+  "workspaces": [
+    "next-visual",
+    "react-visual"
+  ]
 }
