+++ react-visual/src/index.ts
@@ -1,0 +1,4 @@
+import LazyVideo from './LazyVideo'
+import VisualWrapper from './VisualWrapper'
+import { collectDataAttributes } from './lib/attributes'
+export { LazyVideo, VisualWrapper, collectDataAttributes }
