+++ react-visual/tsconfig.pkg.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../tsconfig.build.json",
+  "compilerOptions": {
+    "outDir": "./dist/"
+  },
+  "include": [ "src" ]
+}
