+++ tsconfig.build.json
@@ -1,0 +1,21 @@
+{
+  "compilerOptions": {
+    "target": "es5",
+    "module": "commonjs",
+    "declaration": true,
+    "outDir": "./dist",
+    "strict": true,
+    "jsx": "react-jsx",
+
+    // Required for importing packages, like React, that have `export =`
+    "esModuleInterop": true,
+
+    // Fixes build failures when dependent types have errors, like:
+    // node_modules/next/dist/shared/lib/get-img-props.d.ts:70
+    "skipLibCheck": true
+  },
+  "exclude": [
+    "node_modules",
+    "dist",
+  ]
+}
